requirements.txt
aiohttp==3.7.3
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
boto3==1.15.1
botocore==1.18.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
daphne==3.0.1
defusedxml==0.7.0rc1
Django==3.1.1
django-cors-headers==3.5.0
django-crontab==0.7.1
django-hashers-passlib==0.4
django-rest-swagger==2.2.0
djangorestframework==3.11.1
docutils==0.16
elasticsearch==7.10.1
elasticsearch-dsl==7.3.0
et-xmlfile==1.0.1
geoip2==4.1.0
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.23.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.0.4
hiredis==1.1.0
httplib2==0.19.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
maxminddb==2.0.3
msgpack==1.0.2
multidict==5.1.0
oauth2client==4.1.3
oauthlib==3.1.0
openapi-codec==1.3.2
openpyxl==3.0.5
packaging==20.9
passlib==1.7.2
paymentwall-python==1.0.8
protobuf==3.15.5
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.25.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
s3transfer==0.3.3
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
yarl==1.6.3
zoomus==1.1.3
zope.interface==5.2.0
django-requestlogs==0.2.3
django-ipware==3.0.2
drf_yasg==1.20.0
django-simple-history==3.0.0
PyMySQL==1.0.2
pycryptodome==3.14.1
kafka-python==2.0.2